Chore: rename tile.py file and its classes to better match 3dtiles spec
* tile.py becomes tile_content.py
* TileHeader class becomes TileContentHeader
* TileBody class becomes TileContentBody
* TileType enumeration becomes TileContentType
