+++ docs/api.rst
@@ -9,9 +9,9 @@ Generic Tile
 The py3dtiles module provides some classes to fit into the
 specification:
 
-- *TileContent* with a header *TileHeader* and a body *TileBody*
-- *TileHeader* represents the metadata of the tile (magic value, version, ...)
-- *TileBody* contains varying semantic and geometric data depending on the the tile's type
+- *TileContent* with a header *TileContentHeader* and a body *TileContentBody*
+- *TileContentHeader* represents the metadata of the tile (magic value, version, ...)
+- *TileContentBody* contains varying semantic and geometric data depending on the the tile's type
 
 Moreover, a utility class *TileContentReader* is available to read a tile
 file as well as a simple command line tool to retrieve basic information
@@ -43,12 +43,12 @@ In the current implementation, the *Pnts* class only contains a *FeatureTable*
     >>>
     >>> # tile_content is an instance of the TileContent class
     >>> tile_content
-    <py3dtiles.tile.TileContent>
+    <py3dtiles.tile_content.TileContent>
     >>>
     >>> # extract information about the tile_content header
     >>> th = tile_content.header
     >>> th
-    <py3dtiles.tile.TileHeader>
+    <py3dtiles.tile_content.TileContentHeader>
     >>> th.magic_value
     'pnts'
     >>> th.tile_byte_length
@@ -127,7 +127,7 @@ https://github.com/AnalyticalGraphicsInc/3d-tiles/tree/master/TileFormats/Batche
     >>>
     >>> # tile_content is an instance of the TileContent class
     >>> tile_content
-    <py3dtiles.tile.TileContent>
+    <py3dtiles.tile_content.TileContent>
     >>>
     >>> # extract information about the tile header
     >>> th = tile.header
