+++ py3dtiles/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 from .utils import TileContentReader, convert_to_ecef
-from .tile import TileContent
+from .tile_content import TileContent
 from .feature_table import Feature
 from .gltf import GlTF
 from .pnts import Pnts
